feat: 배경음악 컴포넌트 분리
