+++ src/containers/routine-timer/create-routine/routine-form.tsx
@@ -2,20 +2,17 @@
 
 import { Button } from '@/components/button';
 import { Input } from '@/components/input';
-import React, { useState } from 'react';
+import React from 'react';
 import { useRouter } from 'next/navigation';
-import { Clock, Music } from 'lucide-react';
+import { Clock } from 'lucide-react';
 import { ActivityForm } from './activity-form';
+import { MusicSelector } from '../background-music/music-selector';
 import { useRoutine } from './use-routine';
 import { RouteParams } from './routine-types';
 
 export default function RoutineForm({ params }: RouteParams): JSX.Element {
   const router = useRouter();
   const routineId = params.id;
-  const [youtubeUrl, setYoutubeUrl] = useState('');
-  const [musicVideoId, setMusicVideoId] = useState('');
-  const [isLoadingMusic, setIsLoadingMusic] = useState(false);
-  const [musicError, setMusicError] = useState('');
 
   const {
     routine,
@@ -32,58 +29,8 @@ export default function RoutineForm({ params }: RouteParams): JSX.Element {
     updateRoutineMusic,
   } = useRoutine(routineId);
 
-  const extractVideoId = (url: string): string => {
-    const patterns = [
-      /(?:youtube\.com\/watch\?v=|youtu\.be\/|youtube\.com\/embed\/)([^&\n?#]+)/,
-      /youtube\.com\/watch\?.*v=([^&\n?#]+)/,
-    ];
-
-    const matchResult = patterns.find((pattern) => {
-      const match = url.match(pattern);
-      return match && match[1];
-    });
-
-    if (matchResult) {
-      const match = url.match(matchResult);
-      return match ? match[1] : '';
-    }
-
-    return '';
-  };
-
-  const handleYoutubeUrlChange = async (url: string) => {
-    setYoutubeUrl(url);
-    setMusicError('');
-
-    if (!url.trim()) {
-      setMusicVideoId('');
-      // 음악 데이터 제거
-      updateRoutineMusic('', undefined);
-      return;
-    }
-
-    const videoId = extractVideoId(url);
-
-    if (!videoId) {
-      setMusicError('유튜브 동영상 URL을 다시 확인해주세요.');
-      return;
-    }
-
-    if (videoId === musicVideoId) return;
-
-    setIsLoadingMusic(true);
-    try {
-      setMusicVideoId(videoId);
-      // routine에 음악 데이터 저장
-      updateRoutineMusic(videoId, url);
-    } catch (error) {
-      console.error('Error getting music title:', error);
-      setMusicError('동영상을 찾지 못했어요. 다시 시도해주세요.');
-      setMusicVideoId('');
-      updateRoutineMusic('', undefined);
-    } finally {
-      setIsLoadingMusic(false);
-    }
+  const handleMusicChange = (videoId: string, url?: string) => {
+    updateRoutineMusic(videoId, url);
   };
 
   const handleSaveWithMusic = () => {
@@ -98,16 +45,6 @@ export default function RoutineForm({ params }: RouteParams): JSX.Element {
   const handleRoutineList = () => {
     router.push(`/routine-timer`);
   };
-
-  // 컴포넌트 마운트시 기존 음악 데이터 로드
-  React.useEffect(() => {
-    if (routine.videoId) {
-      setMusicVideoId(routine.videoId);
-      if (routine.url) {
-        setYoutubeUrl(routine.url);
-      }
-    }
-  }, [routine.videoId, routine.url]);
 
   return (
     <div className="p-4 max-w-4xl mx-auto">
@@ -148,40 +85,11 @@ export default function RoutineForm({ params }: RouteParams): JSX.Element {
         </div>
       </div>
 
-      {/* 배경음악 설정 섹션 */}
-      <div className="bg-blue-50 rounded-xl p-6 mb-6">
-        <div className="flex items-center gap-2 mb-4">
-          <Music className="h-5 w-5 text-blue-600" />
-          <h3 className="text-lg font-semibold text-blue-800">배경음악 설정</h3>
-        </div>
-
-        <div className="space-y-3">
-          <Input
-            type="text"
-            value={youtubeUrl}
-            onChange={(e) => handleYoutubeUrlChange(e.target.value)}
-            className="w-full p-3 border rounded-lg"
-            placeholder="유튜브 동영상 URL을 입력하세요 (예: https://www.youtube.com/watch?v=...)"
-          />
-
-          {isLoadingMusic && (
-            <div className="text-blue-600 text-sm">
-              음악 정보를 불러오는 중...
-            </div>
-          )}
-
-          {musicError && (
-            <div className="text-red-500 text-sm">{musicError}</div>
-          )}
-
-          {/* 현재 설정된 음악 표시 */}
-          {routine.videoId && (
-            <div className="text-green-600 text-sm">
-              ✓ 배경음악이 설정되었습니다 (Video ID: {routine.videoId})
-            </div>
-          )}
-        </div>
-      </div>
+      <MusicSelector
+        videoId={routine.videoId || ''}
+        url={routine.url}
+        onMusicChange={handleMusicChange}
+      />
 
       <div className="bg-primary-100 rounded-xl p-6 mb-8">
         <ActivityForm
